backend/scripts/requirements.txt
requests
PyMuPDF
pillow